All `Reader` methods should be `#[inline]`.
As far as performance is concerned, this crate is designed with the
expectation that the compiler will figure out how to elide and coalesce
redundant bounds checks. Hopefully the compiler will avoid constructing a
`Reader` in memory at all in many cases. However, this can only work if
`Reader` methods are inlined.